feat: remove code related to opensearch
