+++ backend/lambdas/utils.ts
@@ -79,15 +79,6 @@ export const decodeUriString = (uriString: string) => {
   }
 };
 
-export const getOpenSearchLambdaConfigOrFail = () => {
-  const getEnv = getGetEnvWithPreassignedContext('Metadata parser lambda');
-  return {
-    openSearchDomain: getEnv('OPENSEARCH_DOMAIN'),
-    region: getEnv('REGION'),
-    metadataIndex: getEnv('METADATA_INDEX'),
-  };
-};
-
 /**
  * Extract and decode key from S3 event record
  */
