+++ frontend/next.config.js
@@ -17,37 +17,14 @@ const nextConfig = {
     unoptimized: true,
   },
   rewrites: () => {
-    const currentMetadataDatabase =
-      process.env.NEXT_PUBLIC_METADATA_DATABASE || '';
-    // path rewrites to only enable "reports" or "graphql_reports" page
-    if (currentMetadataDatabase === 'opensearch') {
-      return {
-        beforeFiles: [
-          {
-            source: '/graphql_reports',
-            destination: '/reports',
-          },
-          {
-            source: '/',
-            destination: '/reports',
-          },
-        ],
-      };
-    } else if (currentMetadataDatabase === 'postgres') {
-      return {
-        beforeFiles: [
-          {
-            source: '/reports',
-            destination: '/graphql_reports',
-          },
-          {
-            source: '/',
-            destination: '/graphql_reports',
-          },
-        ],
-      };
-    }
-    return [];
+    return {
+      beforeFiles: [
+        {
+          source: '/',
+          destination: '/reports',
+        },
+      ],
+    };
   },
 };
