--- frontend/pages/api/file.ts
@@ -1,7 +1,0 @@
-import { NextApiHandler } from 'next';
-
-const handler: NextApiHandler = (req, res) => {
-  res.status(200).json({});
-};
-
-export default handler;
