+++ frontend/shared/__tests__/util.ts
@@ -1,6 +1,6 @@
 import * as F from 'fast-check';
 
-import { fst, snd, prefix, toSearchQueryTerm, toUndefined } from 'shared/util';
+import { prefix, toUndefined } from 'shared/util';
 
 const strOpts: F.StringSharedConstraints = {
   minLength: 1,
@@ -13,18 +13,6 @@ const strOpts: F.StringSharedConstraints = {
 const str1 = () => F.string(strOpts);
 
 describe('Util', () => {
-  test('fst', () => {
-    F.assert(
-      F.property(F.tuple(str1(), str1()), ([a, b]) => fst([a, b]) === a),
-    );
-  });
-
-  test('snd', () => {
-    F.assert(
-      F.property(F.tuple(str1(), str1()), ([a, b]) => snd([a, b]) === b),
-    );
-  });
-
   test('prefix', () => {
     F.assert(
       F.property(str1(), str1(), (a, b) => prefix(a, b) === [a, b].join('')),
@@ -36,27 +24,6 @@ describe('Util', () => {
       F.property(str1(), x => {
         expect(toUndefined(x)).toBeUndefined();
       }),
-    );
-  });
-
-  test('toSearchQueryTerm', () => {
-    F.assert(
-      F.property(str1(), str1(), (key, value) =>
-        expect(toSearchQueryTerm(key, value)).toEqual({
-          match: { [key]: value },
-        }),
-      ),
-    );
-
-    /** It also supports key transformation */
-    F.assert(
-      F.property(str1(), str1(), str1(), (key, value, keyPrefix) =>
-        expect(
-          toSearchQueryTerm(key, value, _key => [keyPrefix, _key].join('')),
-        ).toEqual({
-          match: { [[keyPrefix, key].join('')]: value },
-        }),
-      ),
     );
   });
 });
