+++ frontend/shared/util.ts
@@ -1,14 +1,3 @@
-import saveAs from 'file-saver';
-import { Dispatch, SetStateAction } from 'react';
-
-export function fst<T1, T2>([a, b]: [T1, T2]) {
-  return a;
-}
-
-export function snd<T1, T2>([a, b]: [T1, T2]) {
-  return b;
-}
-
 export const prefix = (p: string, x: string) => [p, x].join('');
 
 /**
@@ -16,27 +5,6 @@ export const prefix = (p: string, x: string) => [p, x].join('');
  */
 export function noop(...args: any[]) {}
 export { noop as toUndefined };
-
-/**
- * Utility for turning a `key` and `value` into an object used for matching said
- * key to value in OpenSearch.
- *
- * Allows optional transformer function for `key`.
- *
- * @param key
- * @param value
- * @param keyFn
- * @returns
- */
-export const toSearchQueryTerm = (
-  key: string,
-  value: string,
-  keyFn?: (x: string) => string,
-) => ({
-  match: {
-    [keyFn ? keyFn(key) : key]: value,
-  },
-});
 
 export const sizeformatter = (size: number | undefined, decimalPlaces = 2) => {
   if (size) {
@@ -54,15 +22,3 @@ export const sizeformatter = (size: number | undefined, decimalPlaces = 2) => {
 
 export const takeOptionValues = (fs: HTMLCollectionOf<HTMLOptionElement>) =>
   Array.from(fs, opt => opt.value as string).filter(Boolean);
-
-export const getKeyAggregations = (size: number | undefined) => {
-  if (!size) return;
-  return {
-    keys: {
-      terms: {
-        field: 'key.keyword',
-        size: size,
-      },
-    },
-  };
-};
