+++ lib/raita-application.ts
@@ -1,11 +1,10 @@
-import * as opensearch from 'aws-cdk-lib/aws-opensearchservice';
 import * as ec2 from 'aws-cdk-lib/aws-ec2';
 import * as iam from 'aws-cdk-lib/aws-iam';
 import { Port } from 'aws-cdk-lib/aws-ec2';
 import { BundlingOutput, NestedStack, NestedStackProps } from 'aws-cdk-lib';
 import { Construct } from 'constructs';
 import { RaitaEnvironment } from './config';
-import { getEnvDependentOsConfiguration, isPermanentStack } from './utils';
+import { isPermanentStack } from './utils';
 
 import { RaitaApiStack } from './raita-api';
 import { DataProcessStack } from './raita-data-process';
@@ -22,7 +21,6 @@ interface ApplicationStackProps extends NestedStackProps {
   readonly jwtTokenIssuer: string;
   readonly vpc: ec2.IVpc;
   readonly raitaSecurityGroup: ec2.ISecurityGroup;
-  readonly openSearchMetadataIndex: string;
   readonly parserConfigurationFile: string;
   readonly sftpPolicyAccountId: string;
   readonly sftpPolicyUserId: string;
@@ -33,9 +31,6 @@ interface ApplicationStackProps extends NestedStackProps {
   readonly cloudfrontDomainName: string;
 }
 
-/**
- * OpenSearch documentation available at: https://docs.aws.amazon.com/opensearch-service/latest/developerguide/what-is.html
- */
 export class ApplicationStack extends NestedStack {
   constructor(scope: Construct, id: string, props: ApplicationStackProps) {
     super(scope, id, props);
@@ -46,7 +41,6 @@ export class ApplicationStack extends NestedStack {
       jwtTokenIssuer,
       vpc,
       raitaSecurityGroup,
-      openSearchMetadataIndex,
       parserConfigurationFile,
       sftpPolicyAccountId,
       sftpPolicyUserId,
@@ -56,14 +50,6 @@ export class ApplicationStack extends NestedStack {
       loramPolicyUserId,
       cloudfrontDomainName,
     } = props;
-
-    // Create and configure OpenSearch domain
-    const openSearchDomain = this.createOpenSearchDomain({
-      name: 'db',
-      raitaEnv,
-      vpc,
-      raitaStackIdentifier,
-    });
 
     // Create a lambda layer containing prisma client and engine, to avoid bundling big engine files for every lambda
     const prismaLambdaLayer = new LayerVersion(this, 'prisma-layer-version', {
@@ -83,8 +69,6 @@ export class ApplicationStack extends NestedStack {
       raitaEnv,
       stackId,
       vpc,
-      openSearchDomain: openSearchDomain,
-      openSearchMetadataIndex: openSearchMetadataIndex,
       parserConfigurationFile: parserConfigurationFile,
       sftpPolicyAccountId: sftpPolicyAccountId,
       sftpPolicyUserId: sftpPolicyUserId,
@@ -100,12 +84,10 @@ export class ApplicationStack extends NestedStack {
       inspectionDataBucket: dataProcessStack.inspectionDataBucket,
       dataReceptionBucket: dataProcessStack.dataReceptionBucket,
       csvDataBucket: dataProcessStack.csvDataBucket,
-      openSearchDomain: openSearchDomain,
       raitaEnv,
       stackId,
       jwtTokenIssuer,
       raitaStackIdentifier: raitaStackIdentifier,
-      openSearchMetadataIndex: openSearchMetadataIndex,
       cloudfrontDomainName: cloudfrontDomainName,
       vpc,
       raitaSecurityGroup,
@@ -119,7 +101,6 @@ export class ApplicationStack extends NestedStack {
         vpc,
         securityGroup: raitaSecurityGroup,
         albDns: raitaApiStack.alb.loadBalancerDnsName,
-        databaseDomainEndpoint: openSearchDomain.domainEndpoint,
       });
     }
 
@@ -135,16 +116,6 @@ export class ApplicationStack extends NestedStack {
       });
     }
 
-    // Grant data processor lambdas permissions to call OpenSearch endpoints
-    // TODO: RAITA-273 Review if this can be dropped as permissions are given directly to metadata index
-    this.createManagedPolicy({
-      name: 'DataProcessOpenSearchHttpPolicy',
-      raitaStackIdentifier,
-      serviceRoles: [dataProcessStack.dataProcessorLambdaServiceRole],
-      resources: [openSearchDomain.domainArn],
-      actions: ['es:ESHttpGet', 'es:ESHttpPost', 'es:ESHttpPut'],
-    });
-
     // Grant api lambdas permissions to get API-key from
     // SSM Parameterstore
     this.createManagedPolicy({
@@ -159,19 +130,6 @@ export class ApplicationStack extends NestedStack {
         `arn:aws:ssm:${this.region}:${this.account}:parameter/${SSM_API_KEY}`,
       ],
       actions: ['ssm:GetParameter'],
-    });
-
-    // Grant api lambdas permissions to call OpenSearch endpoints
-    // TODO: RAITA-273 Review if this can be dropped as permissions are given directly to metadata index
-    this.createManagedPolicy({
-      name: 'ApiOpenSearchHttpPolicy',
-      raitaStackIdentifier,
-      serviceRoles: [
-        raitaApiStack.raitaApiLambdaServiceRole,
-        raitaApiStack.raitaApiDeleteRequestLambdaServiceRole,
-      ],
-      resources: [openSearchDomain.domainArn],
-      actions: ['es:ESHttpGet', 'es:ESHttpPost', 'es:ESHttpPut'],
     });
 
     // Grant api handleZipRequestLambda permission to invoke the
@@ -190,84 +148,6 @@ export class ApplicationStack extends NestedStack {
       serviceRoles: [raitaApiStack.raitaApiGraphqlLambdaServiceRole],
       resources: [raitaApiStack.handleCsvGenerationFn.functionArn],
       actions: ['lambda:invokeFunction'],
-    });
-
-    // Allow connections to OpenSearch
-    openSearchDomain.connections.allowFrom(
-      raitaSecurityGroup,
-      Port.tcp(443),
-      'Allow connections to Opensearch from Raita Security group.',
-    );
-
-    // TODO: This can be likely simplified by assigning Default Raita
-    // serity group to below lambdas - allowFrom calls should become unnecessary
-    // (Jira 207)
-    // Allow traffic from lambdas to OpenSearch
-    openSearchDomain.connections.allowFrom(
-      dataProcessStack.handleInspectionFileEventFn,
-      Port.allTraffic(),
-      'Allows parser lambda to connect to Opensearch.',
-    );
-    openSearchDomain.connections.allowFrom(
-      raitaApiStack.handleFilesRequestFn,
-      Port.allTraffic(),
-      'Allows parser lambda to connect to Opensearch.',
-    );
-    openSearchDomain.connections.allowFrom(
-      raitaApiStack.handleMetaRequestFn,
-      Port.allTraffic(),
-      'Allows meta endpoint handler lambda to connect to Opensearch.',
-    );
-    openSearchDomain.connections.allowFrom(
-      raitaApiStack.handleDeleteRequestFn,
-      Port.allTraffic(),
-      'Allows delete endpoint handler lambda to connect to Opensearch.',
-    );
-  }
-
-  /**
-   * Creates OpenSearch domain
-   */
-  private createOpenSearchDomain({
-    name,
-    raitaEnv,
-    vpc,
-    raitaStackIdentifier,
-  }: {
-    name: string;
-    raitaEnv: RaitaEnvironment;
-    vpc: ec2.IVpc;
-    raitaStackIdentifier: string;
-  }) {
-    const domainName = `${name}-${raitaStackIdentifier}`;
-    const domainArn =
-      'arn:aws:es:' +
-      this.region +
-      ':' +
-      this.account +
-      ':domain/' +
-      domainName +
-      '/*';
-    return new opensearch.Domain(this, domainName, {
-      domainName,
-      version: opensearch.EngineVersion.OPENSEARCH_1_3,
-      enableVersionUpgrade: true,
-      nodeToNodeEncryption: true,
-      enforceHttps: true,
-      vpc,
-      encryptionAtRest: {
-        enabled: true,
-      },
-      ...getEnvDependentOsConfiguration(raitaEnv, vpc.privateSubnets),
-      // See RAITA-231 for removing wildcard permissions from AnyPrincipal
-      accessPolicies: [
-        new iam.PolicyStatement({
-          effect: iam.Effect.ALLOW,
-          actions: ['es:ESHttp*'],
-          principals: [new iam.AnyPrincipal()],
-          resources: [domainArn],
-        }),
-      ],
     });
   }
