+++ lib/utils.ts
@@ -1,6 +1,5 @@
 import { RemovalPolicy, aws_lambda_nodejs } from 'aws-cdk-lib';
 import { EbsDeviceVolumeType, ISubnet } from 'aws-cdk-lib/aws-ec2';
-import { DomainProps } from 'aws-cdk-lib/aws-opensearchservice';
 import * as path from 'path';
 import { readFileSync } from 'fs';
 import {
@@ -49,62 +48,6 @@ export const isProductionStack = (
 export const isPermanentStack = (stackId: string, raitaEnv: RaitaEnvironment) =>
   isDevelopmentMainStack(stackId, raitaEnv) ||
   isProductionStack(stackId, raitaEnv);
-
-/**
- * Returns env dependent OpenSearch configuration properties
- */
-export const getEnvDependentOsConfiguration = (
-  env: RaitaEnvironment,
-  subnets: ISubnet[],
-) => {
-  const envDependentProperties: Record<
-    RaitaEnvironment,
-    Partial<DomainProps>
-  > = {
-    prod: {
-      removalPolicy: getRemovalPolicy(env),
-      ebs: {
-        volumeSize: 200,
-        volumeType: EbsDeviceVolumeType.GENERAL_PURPOSE_SSD_GP3,
-      },
-      capacity: {
-        masterNodes: 3,
-        masterNodeInstanceType: 'm6g.large.search',
-        dataNodes: 2,
-        dataNodeInstanceType: 'm6g.large.search',
-      },
-      // Must be enabled if VPC contains multiple private subnets.
-      zoneAwareness: {
-        enabled: true,
-        availabilityZoneCount: subnets.length,
-      },
-      vpcSubnets: [
-        {
-          subnets: subnets,
-        },
-      ],
-    },
-    // identical conf in dev to accurately test with large amount of data
-    // this can be changed back to old conf (in comments) if needed
-    dev: {
-      removalPolicy: getRemovalPolicy(env),
-      ebs: {
-        volumeSize: 50,
-        volumeType: EbsDeviceVolumeType.GENERAL_PURPOSE_SSD,
-      },
-      capacity: {
-        dataNodes: 1,
-        dataNodeInstanceType: 't3.small.search',
-      },
-      vpcSubnets: [
-        {
-          subnets: subnets.slice(0, 1),
-        },
-      ],
-    },
-  };
-  return envDependentProperties[env];
-};
 
 /**
  * Get environment variables needed to log into the database.
