+++ package.json
@@ -56,19 +56,16 @@
     "@as-integrations/aws-lambda": "^3.1.0",
     "@aws-sdk/client-ecs": "^3.470.0",
     "@aws-sdk/client-lambda": "^3.470.0",
-    "@aws-sdk/client-opensearch": "^3.470.0",
     "@aws-sdk/client-s3": "^3.470.0",
     "@aws-sdk/client-sqs": "^3.609.0",
     "@aws-sdk/credential-provider-node": "^3.470.0",
     "@aws-sdk/lib-storage": "^3.470.0",
     "@aws-sdk/s3-request-presigner": "^3.470.0",
-    "@opensearch-project/opensearch": "^2.4.0",
     "@prisma/client": "^5.12.1",
     "@side/jest-runtime": "^1.1.0",
     "archiver": "^6.0.1",
     "aws-cdk-lib": "^2.148.1",
     "aws-lambda": "^1.0.7",
-    "aws-opensearch-connector": "^1.1.0",
     "axios": "^1.6.2",
     "cloneable-readable": "^3.0.0",
     "constructs": "^10.3.0",
@@ -79,7 +76,6 @@
     "jsonwebtoken": "^9.0.2",
     "jszip": "^3.10.1",
     "jwk-to-pem": "^2.0.5",
-    "opensearch": "^0.0.3",
     "pino": "^8.16.2",
     "pino-lambda": "^4.4.0",
     "postgres": "^3.4.3",
